+++ database/sql/src/main/scala/cromwell/database/slick/tables/MetadataEntryComponent.scala
@@ -62,6 +62,13 @@ trait MetadataEntryComponent {
     } yield metadataEntry).sortBy(_.metadataTimestamp)
   )
 
+  val metadataEntriesForWorkflowSortedById = Compiled(
+    (workflowExecutionUuid: Rep[String]) => (for {
+      metadataEntry <- metadataEntries
+      if metadataEntry.workflowExecutionUuid === workflowExecutionUuid
+    } yield metadataEntry).sortBy(_.metadataEntryId)
+  )
+
   val countMetadataEntriesForWorkflowExecutionUuid = Compiled(
     (rootWorkflowId: Rep[String], expandSubWorkflows: Rep[Boolean]) => {
       val targetWorkflowIds = for {
@@ -75,22 +82,6 @@ trait MetadataEntryComponent {
         if metadata.workflowExecutionUuid in targetWorkflowIds // Uses `METADATA_WORKFLOW_IDX`
       } yield metadata
     }.size
-  )
-
-  val metadataEntriesWithoutLabelsForRootWorkflowId = Compiled(
-    (rootWorkflowId: Rep[String]) => {
-      val targetWorkflowIds = for {
-        summary <- workflowMetadataSummaryEntries
-        // Uses `IX_WORKFLOW_METADATA_SUMMARY_ENTRY_RWEU`, `UC_WORKFLOW_METADATA_SUMMARY_ENTRY_WEU`
-        if summary.rootWorkflowExecutionUuid === rootWorkflowId || summary.workflowExecutionUuid === rootWorkflowId
-      } yield summary.workflowExecutionUuid
-
-      for {
-        metadata <- metadataEntries
-        if metadata.workflowExecutionUuid in targetWorkflowIds // Uses `METADATA_WORKFLOW_IDX`
-        if !(metadata.metadataKey like "labels:%")
-      } yield metadata
-    }
   )
 
   val metadataEntryExistsForWorkflowExecutionUuid = Compiled(