+++ engine/src/main/scala/cromwell/webservice/routes/MetadataRouteSupport.scala
@@ -1,5 +1,8 @@
 package cromwell.webservice.routes
 
+import java.time.{OffsetDateTime, Duration => JDuration}
+import java.util.concurrent.TimeUnit
+
 import akka.actor.{ActorRef, ActorRefFactory}
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport._
 import akka.http.scaladsl.marshalling.ToResponseMarshallable
@@ -11,19 +14,27 @@ import akka.util.Timeout
 import cats.data.NonEmptyList
 import cats.data.Validated.{Invalid, Valid}
 import cromwell.core.Dispatcher.ApiDispatcher
+import cromwell.core.instrumentation.InstrumentationPrefixes.ServicesPrefix
 import cromwell.core.labels.Labels
-import cromwell.core.{WorkflowId, path => _}
+import cromwell.core.{WorkflowId, WorkflowMetadataKeys, path => _}
 import cromwell.engine.instrumentation.HttpInstrumentation
 import cromwell.server.CromwellShutdown
 import cromwell.services._
+import cromwell.services.instrumentation.{CromwellBucket, CromwellIncrement, CromwellTiming}
+import cromwell.services.instrumentation.InstrumentationService.InstrumentationServiceMessage
+import cromwell.services.metadata.MetadataArchiveStatus
 import cromwell.services.metadata.MetadataService._
+import cromwell.services.metadata.impl.MetadataServiceActor
 import cromwell.webservice.LabelsManagerActor
 import cromwell.webservice.LabelsManagerActor._
 import cromwell.webservice.WebServiceUtils.EnhancedThrowable
 import cromwell.webservice.WorkflowJsonSupport._
 import cromwell.webservice.routes.CromwellApiService._
 import cromwell.webservice.routes.MetadataRouteSupport._
+import cromwell.webservice.WebServiceUtils._
+import spray.json.{JsObject, JsString}
 
+import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Future, TimeoutException}
 import scala.util.{Failure, Success}
 
@@ -85,19 +96,7 @@ trait MetadataRouteSupport extends HttpInstrumentation {
           entity(as[Map[String, String]]) { parameterMap =>
             instrumentRequest {
               Labels.validateMapOfLabels(parameterMap) match {
-                case Valid(labels) =>
-                  val response = validateWorkflowIdInMetadataSummaries(possibleWorkflowId, serviceRegistryActor) flatMap { id =>
-                    val lma = actorRefFactory.actorOf(LabelsManagerActor.props(serviceRegistryActor).withDispatcher(ApiDispatcher))
-                    lma.ask(LabelsAddition(LabelsData(id, labels))).mapTo[LabelsManagerActorResponse]
-                  }
-                  onComplete(response) {
-                    case Success(r: BuiltLabelsManagerResponse) => complete(r.response)
-                    case Success(e: FailedLabelsManagerResponse) => e.reason.failRequest(StatusCodes.InternalServerError)
-                    case Failure(e: UnrecognizedWorkflowException) => e.failRequest(StatusCodes.NotFound)
-                    case Failure(e: TimeoutException) => e.failRequest(StatusCodes.ServiceUnavailable)
-                    case Failure(e) => e.errorRequest(StatusCodes.InternalServerError)
-
-                  }
+                case Valid(labels) => patchLabelsRequest(possibleWorkflowId, labels, serviceRegistryActor, actorRefFactory)
                 case Invalid(e) =>
                   val iae = new IllegalArgumentException(e.toList.mkString(","))
                   iae.failRequest(StatusCodes.BadRequest)
@@ -127,6 +126,27 @@ trait MetadataRouteSupport extends HttpInstrumentation {
 }
 
 object MetadataRouteSupport {
+
+  private def processMetadataArchivedResponse(workflowId: WorkflowId,
+                                              archiveStatus: MetadataArchiveStatus,
+                                              endTime: Option[OffsetDateTime],
+                                              additionalMsg: String = ""): JsObject = {
+    val baseMessage = "Cromwell has archived this workflow's metadata according to the lifecycle policy."
+    val timeSinceMessage = endTime map { timestamp =>
+      val duration = FiniteDuration(JDuration.between(timestamp, OffsetDateTime.now()).toMillis, TimeUnit.MILLISECONDS)
+      s" The workflow completed at $timestamp, which was ${duration} ago."
+    } getOrElse("")
+    val additionalDetails = if (archiveStatus == MetadataArchiveStatus.ArchivedAndDeleted)
+      " It is available in the archive bucket, or via a support request in the case of a managed instance."
+    else ""
+
+    JsObject(Map(
+      WorkflowMetadataKeys.Id -> JsString(workflowId.toString),
+      WorkflowMetadataKeys.MetadataArchiveStatus -> JsString(archiveStatus.toString),
+      WorkflowMetadataKeys.Message -> JsString(baseMessage + timeSinceMessage + additionalDetails + additionalMsg)
+    ))
+  }
+
   def metadataLookup(possibleWorkflowId: String,
                      request: WorkflowId => BuildMetadataJsonAction,
                      serviceRegistryActor: ActorRef)
@@ -145,7 +165,50 @@ object MetadataRouteSupport {
                                   serviceRegistryActor: ActorRef)
                                  (implicit timeout: Timeout,
                                   ec: ExecutionContext): Future[MetadataJsonResponse] = {
-    validateWorkflowIdInMetadata(possibleWorkflowId, serviceRegistryActor) flatMap { w => serviceRegistryActor.ask(request(w)).mapTo[MetadataJsonResponse] }
+
+    def recordHistoricalMetadataLookupMetrics(endTime: Option[OffsetDateTime]): Unit = {
+      val timeSinceEndTime = endTime match {
+        case Some(et) => FiniteDuration(JDuration.between(et, OffsetDateTime.now()).toMillis, TimeUnit.MILLISECONDS)
+        case None => 0.seconds
+      }
+
+      val lagInstrumentationPath = MetadataServiceActor.MetadataInstrumentationPrefix :+ "archiver" :+ "historical_metadata_lookup" :+ "lag"
+      val lagMessage = InstrumentationServiceMessage(CromwellTiming(CromwellBucket(ServicesPrefix.toList, lagInstrumentationPath), timeSinceEndTime))
+      serviceRegistryActor ! lagMessage
+
+      val interestingDayMarks = (5.to(55, step = 5) ++ 0.to(4)).map(d => (d.days, s"${d}_day_old"))
+      val interestingMonthMarks = 2.to(11).map(m => (30.days * m.longValue(), s"${m}_month_old"))
+
+      (interestingDayMarks ++ interestingMonthMarks) foreach {
+        case (timeSpan, metricName) if timeSinceEndTime >= timeSpan =>
+          val oldMetadataCounterPath = MetadataServiceActor.MetadataInstrumentationPrefix :+ "archiver" :+ "historical_metadata_lookup" :+ "lookup_age_counts" :+ metricName
+          val oldMetadataCounterMessage = InstrumentationServiceMessage(CromwellIncrement(CromwellBucket(ServicesPrefix.toList, oldMetadataCounterPath)))
+          serviceRegistryActor ! oldMetadataCounterMessage
+        case _ => // Do nothing
+      }
+    }
+
+    def checkIfMetadataDeletedAndRespond(id: WorkflowId,
+                                         metadataRequest: BuildWorkflowMetadataJsonWithOverridableSourceAction): Future[MetadataJsonResponse] = {
+      serviceRegistryActor.ask(FetchWorkflowMetadataArchiveStatusAndEndTime(id)).mapTo[FetchWorkflowArchiveStatusAndEndTimeResponse] flatMap {
+        case WorkflowMetadataArchivedStatusAndEndTime(archiveStatus, endTime) =>
+          recordHistoricalMetadataLookupMetrics(endTime)
+          if (archiveStatus.isDeleted) Future.successful(SuccessfulMetadataJsonResponse(metadataRequest, processMetadataArchivedResponse(id, archiveStatus, endTime)))
+          else serviceRegistryActor.ask(request(id)).mapTo[MetadataJsonResponse]
+        case FailedToGetArchiveStatusAndEndTime(e) => Future.failed(e)
+      }
+    }
+
+    validateWorkflowIdInMetadata(possibleWorkflowId, serviceRegistryActor) flatMap { id =>
+      /*
+        for requests made to one of /metadata, /logs or /outputs endpoints, perform an additional check to see
+        if metadata for the workflow has been archived and deleted or not (as they interact with metadata table)
+      */
+      request(id) match {
+        case m: BuildWorkflowMetadataJsonWithOverridableSourceAction => checkIfMetadataDeletedAndRespond(id, m)
+        case _ => serviceRegistryActor.ask(request(id)).mapTo[MetadataJsonResponse]
+      }
+    }
   }
 
   def completeMetadataBuilderResponse(response: Future[MetadataJsonResponse]): Route = {
@@ -175,5 +238,46 @@ object MetadataRouteSupport {
       case Failure(e: TimeoutException) => e.failRequest(StatusCodes.ServiceUnavailable)
       case Failure(e) => e.errorRequest(StatusCodes.InternalServerError)
     }
+  }
+
+  def completePatchLabelsResponse(response: Future[LabelsManagerActorResponse]): Route = {
+    onComplete(response) {
+      case Success(r: BuiltLabelsManagerResponse) => complete(r.response)
+      case Success(r: WorkflowArchivedLabelsManagerResponse) => completeResponse(StatusCodes.BadRequest, r.response, Seq.empty)
+      case Success(e: FailedLabelsManagerResponse) => e.reason.failRequest(StatusCodes.InternalServerError)
+      case Failure(e: UnrecognizedWorkflowException) => e.failRequest(StatusCodes.NotFound)
+      case Failure(e: TimeoutException) => e.failRequest(StatusCodes.ServiceUnavailable)
+      case Failure(e) => e.errorRequest(StatusCodes.InternalServerError)
+    }
+  }
+
+  def patchLabelsRequest(possibleWorkflowId: String,
+                         labels: Labels,
+                         serviceRegistryActor: ActorRef,
+                         actorRefFactory: ActorRefFactory)
+                        (implicit timeout: Timeout, ec: ExecutionContext): Route = {
+
+    def checkIfMetadataArchivedAndRespond(id: WorkflowId, archiveStatusResponse: FetchWorkflowArchiveStatusAndEndTimeResponse): Future[LabelsManagerActorResponse] = {
+      archiveStatusResponse match {
+        case WorkflowMetadataArchivedStatusAndEndTime(archiveStatus, endTime) =>
+          if (archiveStatus.isArchived) {
+            val message = " As a result, new labels can't be added or existing labels can't be updated for this workflow."
+            Future.successful(WorkflowArchivedLabelsManagerResponse(processMetadataArchivedResponse(id, archiveStatus, endTime, message)))
+          }
+          else {
+            val lma = actorRefFactory.actorOf(LabelsManagerActor.props(serviceRegistryActor).withDispatcher(ApiDispatcher))
+            lma.ask(LabelsAddition(LabelsData(id, labels))).mapTo[LabelsManagerActorResponse]
+          }
+        case FailedToGetArchiveStatusAndEndTime(e) => Future.failed(e)
+      }
+    }
+
+    val response = for {
+      id <- validateWorkflowIdInMetadataSummaries(possibleWorkflowId, serviceRegistryActor)
+      archiveStatusResponse <- serviceRegistryActor.ask(FetchWorkflowMetadataArchiveStatusAndEndTime(id)).mapTo[FetchWorkflowArchiveStatusAndEndTimeResponse]
+      response <- checkIfMetadataArchivedAndRespond(id, archiveStatusResponse)
+    } yield response
+
+    completePatchLabelsResponse(response)
   }
 }