+++ project/Settings.scala
@@ -1,7 +1,7 @@
 import ContinuousIntegration._
 import Dependencies._
 import GenerateRestApiDocs._
-import Merging.customMergeStrategy
+import Merging._
 import Publishing._
 import Testing._
 import Version._
@@ -12,13 +12,6 @@ import sbtassembly.AssemblyPlugin.autoImport._
 import sbtdocker.{DockerPlugin, Instruction, Instructions}
 
 object Settings {
-
-  val commonResolvers = List(
-    Resolver.jcenterRepo,
-    "Broad Artifactory Releases" at "https://broadinstitute.jfrog.io/broadinstitute/libs-release/",
-    "Broad Artifactory Snapshots" at "https://broadinstitute.jfrog.io/broadinstitute/libs-snapshot/",
-    Resolver.sonatypeRepo("releases")
-  )
 
   /* The reason why -Xmax-classfile-name is set is because this will fail
      to build on Docker otherwise.  The reason why it's 200 is because it
@@ -95,12 +88,12 @@ object Settings {
 
   // 2.12.13 blocked on the release of sbt-scoverage 1.6.2 https://github.com/scoverage/sbt-scoverage/issues/319
   val Scala2_12Version = "2.12.12"
-  val ScalaVersion = Scala2_12Version
-  val sharedSettings =
-    cromwellVersionWithGit ++ artifactorySettings ++ List(
+  val ScalaVersion: String = Scala2_12Version
+  val sharedSettings: Seq[Setting[_]] =
+    cromwellVersionWithGit ++ publishingSettings ++ List(
     organization := "org.broadinstitute",
     scalaVersion := ScalaVersion,
-    resolvers ++= commonResolvers,
+    resolvers ++= additionalResolvers,
     // Don't run tasks in parallel, especially helps in low CPU environments like Travis
     Global / parallelExecution := false,
     Global / concurrentRestrictions ++= List(
@@ -131,7 +124,7 @@ object Settings {
         - https://cromwell.readthedocs.io/en/develop/backends/Google/#issues-with-composite-files
         - https://cloud.google.com/storage/docs/gsutil/addlhelp/CRC32CandInstallingcrcmod
    */
-  val installGcloudSettings: List[Def.Setting[Seq[Instruction]]] = List(
+  val installGcloudSettings: List[Setting[Seq[Instruction]]] = List(
     dockerCustomSettings := List(
       Instructions.Env("PATH", "$PATH:/usr/local/gcloud/google-cloud-sdk/bin"),
       // instructions to install `crcmod`
@@ -147,15 +140,15 @@ object Settings {
       Instructions.Run("curl https://dl.google.com/dl/cloudsdk/release/google-cloud-sdk.tar.gz > /tmp/google-cloud-sdk.tar.gz"),
       Instructions.Run("""mkdir -p /usr/local/gcloud \
                          | && tar -C /usr/local/gcloud -xvf /tmp/google-cloud-sdk.tar.gz \
-                         | && /usr/local/gcloud/google-cloud-sdk/install.sh""".stripMargin),
+                         | && /usr/local/gcloud/google-cloud-sdk/install.sh""".stripMargin)
     )
   )
 
   val swaggerUiSettings = List(Compile / resourceGenerators += writeSwaggerUiVersionConf)
   val backendSettings = List(addCompilerPlugin(kindProjectorPlugin))
-  val engineSettings = swaggerUiSettings
-  val cromiamSettings = swaggerUiSettings
-  val drsLocalizerSettings = installGcloudSettings
+  val engineSettings: List[Setting[_]] = swaggerUiSettings
+  val cromiamSettings: List[Setting[_]] = swaggerUiSettings
+  val drsLocalizerSettings: List[Setting[_]] = installGcloudSettings
 
   private def buildProject(project: Project,
                            projectName: String,
@@ -228,7 +221,7 @@ object Settings {
           .settings(publish := {})
           .settings(generateRestApiDocsSettings)
           .settings(ciSettings)
-          .settings(rootArtifactorySettings)
+          .settings(rootPublishingSettings)
       )
 
       buildProject(project, "root", Nil, builders)