+++ services/src/main/scala/cromwell/services/metadata/impl/archiver/ArchiveMetadataConfig.scala
@@ -6,8 +6,8 @@ import common.Checked
 import common.validation.Validation._
 import cromwell.core.filesystem.CromwellFileSystems
 import cromwell.core.path.PathFactory.PathBuilders
-import cromwell.core.path.{Path, PathBuilderFactory, PathFactory}
-import cromwell.core.{WorkflowId, WorkflowOptions}
+import cromwell.core.path.PathBuilderFactory
+import cromwell.core.WorkflowOptions
 import net.ceedubs.ficus.Ficus._
 
 import scala.concurrent.Await
@@ -17,23 +17,20 @@ import scala.util.Try
 
 final case class ArchiveMetadataConfig(pathBuilders: PathBuilders,
                                        bucket: String,
-                                       interval: FiniteDuration,
-                                       databaseStreamFetchSize: Int,
+                                       backoffInterval: FiniteDuration,
                                        archiveDelay: FiniteDuration,
-                                       debugLogging: Boolean) {
-  def makePath(workflowId: WorkflowId): Path = PathFactory.buildPath(ArchiveMetadataConfig.pathForWorkflow(workflowId, bucket), pathBuilders)
+                                       instrumentationInterval: FiniteDuration,
+                                       debugLogging: Boolean,
+                                       batchSize: Long) {
 }
 
 object ArchiveMetadataConfig {
-
-  // TODO: Confirm if this makes sense to the users? Should we store /bucket/parent-wf-id/sub-wf-id or /bucket/wf-id ?
-  // When deciding keep in mind that workflows can nest to arbitrary depth, mustn't exceed path length limits: /bucket/parent-wf-id/parent-wf-id/parent-wf-id/parent-wf-id/sub-wf-id
-  def pathForWorkflow(id: WorkflowId, bucket: String) = s"gs://$bucket/$id/$id.csv"
-
   def parseConfig(archiveMetadataConfig: Config)(implicit system: ActorSystem): Checked[ArchiveMetadataConfig] = {
     val defaultMaxInterval: FiniteDuration = 5 minutes
     val defaultArchiveDelay = 365 days
+    val defaultInstrumentationInterval = 1 minute
     val defaultDebugLogging = true
+    val defaultBatchSize: Long = 1
 
     for {
       _ <- Try(archiveMetadataConfig.getConfig("filesystems.gcs")).toCheckedWithContext("parse archiver 'filesystems.gcs' field from config")
@@ -41,10 +38,11 @@ object ArchiveMetadataConfig {
       pathBuilders <- Try(Await.result(PathBuilderFactory.instantiatePathBuilders(pathBuilderFactories.values.toList, WorkflowOptions.empty), 60.seconds))
         .toCheckedWithContext("construct archiver path builders from factories")
       bucket <- Try(archiveMetadataConfig.getString("bucket")).toCheckedWithContext("parse Carboniter 'bucket' field from config")
-      interval <- Try(archiveMetadataConfig.getOrElse[FiniteDuration]("interval", defaultMaxInterval)).toChecked
-      databaseStreamFetchSize <- Try(archiveMetadataConfig.getOrElse[Int]("database-stream-fetch-size", 100)).toChecked
+      backoffInterval <- Try(archiveMetadataConfig.getOrElse[FiniteDuration]("backoff-interval", defaultMaxInterval)).toChecked
       archiveDelay <- Try(archiveMetadataConfig.getOrElse("archive-delay", defaultArchiveDelay)).toChecked
+      instrumentationInterval <- Try(archiveMetadataConfig.getOrElse("instrumentation-interval", defaultInstrumentationInterval)).toChecked
       debugLogging <- Try(archiveMetadataConfig.getOrElse("debug-logging", defaultDebugLogging)).toChecked
-    } yield ArchiveMetadataConfig(pathBuilders, bucket, interval, databaseStreamFetchSize, archiveDelay, debugLogging)
+      batchSize <- Try(archiveMetadataConfig.getOrElse("batch-size", defaultBatchSize)).toChecked
+    } yield ArchiveMetadataConfig(pathBuilders, bucket, backoffInterval, archiveDelay, instrumentationInterval, debugLogging, batchSize)
   }
 }